requirements.txt
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
binance-connector>=3.6.0
tenacity>=8.2.3
tqdm>=4.66.0
matplotlib>=3.8.0
requests>=2.31.0
pytest>=8.0.0
